some refactoring + cleaner parameter & file picking
